+++ PgRoutiner/Builder/DiffBuilder/PgDiffBuilder.cs
@@ -44,7 +44,8 @@ public partial class PgDiffBuilder : CodeBuilders.Code
         this.title = title;
         this.targetBuilder = targetBuilder;
 
-        var ste = source.GetTables(new Settings { Schema = settings.Schema }, skipSimilar: settings.DiffSkipSimilarTo);
+        var diffSettings = new Settings { SchemaSimilarTo = settings.SchemaSimilarTo, SchemaNotSimilarTo = settings.SchemaNotSimilarTo };
+        var ste = source.GetTables(diffSettings, skipSimilar: settings.DiffSkipSimilarTo);
         this.sourceTables = ste
             .Where(t => t.Type == PgType.Table)
             .ToDictionary(t => new Table(t.Schema, t.Name), t => t);
@@ -52,24 +53,24 @@ public partial class PgDiffBuilder : CodeBuilders.Code
             .Where(t => t.Type == PgType.View)
             .ToDictionary(t => new Table(t.Schema, t.Name), t => t);
         this.sourceRoutines = source
-            .GetRoutineGroups(new Settings { Schema = settings.Schema }, skipSimilar: settings.DiffSkipSimilarTo)
+            .GetRoutineGroups(diffSettings, skipSimilar: settings.DiffSkipSimilarTo)
             .SelectMany(g => g)
             .ToDictionary(r => new Routine(r.SpecificSchema,
                 r.RoutineName,
             $"({string.Join(", ", r.Parameters.Select(p => $"{p.Name} {p.DataType}{(p.IsArray ? "[]" : "")}"))})"),
                 r => r);
-        this.sourceDomains = source.GetDomains(new Settings { Schema = settings.Schema }, skipSimilar: settings.DiffSkipSimilarTo)
+        this.sourceDomains = source.GetDomains(diffSettings, skipSimilar: settings.DiffSkipSimilarTo)
             .ToDictionary(t => new Domain(t.Schema, t.Name), t => t);
         this._sourceLines = sourceBuilder.GetRawRoutinesDumpLines(settings.DiffPrivileges, out var sourceTypes);
-        this.sourceTypes = source.FilterTypes(sourceTypes, new Settings { Schema = settings.Schema }, skipSimilar: settings.DiffSkipSimilarTo)
+        this.sourceTypes = source.FilterTypes(sourceTypes, diffSettings, skipSimilar: settings.DiffSkipSimilarTo)
             .ToDictionary(t => new Type(t.Schema, t.Name), t => t);
-        this.sourceSchemas = source.GetSchemas(new Settings { Schema = settings.Schema }, skipSimilar: settings.DiffSkipSimilarTo)
+        this.sourceSchemas = source.GetSchemas(diffSettings, skipSimilar: settings.DiffSkipSimilarTo)
             .Where(s => !string.Equals(s, "public"))
             .ToHashSet();
-        this.sourceSeqs = source.GetSequences(new Settings { Schema = settings.Schema }, skipSimilar: settings.DiffSkipSimilarTo)
+        this.sourceSeqs = source.GetSequences(diffSettings, skipSimilar: settings.DiffSkipSimilarTo)
             .ToDictionary(t => new Seq(t.Schema, t.Name), t => t);
 
-        var tte = target.GetTables(new Settings { Schema = settings.Schema }, skipSimilar: settings.DiffSkipSimilarTo);
+        var tte = target.GetTables(diffSettings, skipSimilar: settings.DiffSkipSimilarTo);
         this.targetTables = tte
             .Where(t => t.Type == PgType.Table)
             .ToDictionary(t => new Table(t.Schema, t.Name), t => t);
@@ -77,21 +78,21 @@ public partial class PgDiffBuilder : CodeBuilders.Code
             .Where(t => t.Type == PgType.View)
             .ToDictionary(t => new Table(t.Schema, t.Name), t => t);
         this.targetRoutines = target
-            .GetRoutineGroups(new Settings { Schema = settings.Schema }, skipSimilar: settings.DiffSkipSimilarTo)
+            .GetRoutineGroups(diffSettings, skipSimilar: settings.DiffSkipSimilarTo)
             .SelectMany(g => g)
             .ToDictionary(r => new Routine(r.SpecificSchema,
                 r.RoutineName,
             $"({string.Join(", ", r.Parameters.Select(p => $"{p.Name} {p.DataType}{(p.IsArray ? "[]" : "")}"))})"),
                 r => r);
-        this.targetDomains = target.GetDomains(new Settings { Schema = settings.Schema }, skipSimilar: settings.DiffSkipSimilarTo)
+        this.targetDomains = target.GetDomains(diffSettings, skipSimilar: settings.DiffSkipSimilarTo)
             .ToDictionary(t => new Domain(t.Schema, t.Name), t => t);
         this._targetLines = targetBuilder.GetRawRoutinesDumpLines(settings.DiffPrivileges, out var targetTypes);
-        this.targetTypes = target.FilterTypes(targetTypes, new Settings { Schema = settings.Schema }, skipSimilar: settings.DiffSkipSimilarTo)
+        this.targetTypes = target.FilterTypes(targetTypes, diffSettings, skipSimilar: settings.DiffSkipSimilarTo)
             .ToDictionary(t => new Type(t.Schema, t.Name), t => t);
-        this.targetSchemas = target.GetSchemas(new Settings { Schema = settings.Schema }, skipSimilar: settings.DiffSkipSimilarTo)
+        this.targetSchemas = target.GetSchemas(diffSettings, skipSimilar: settings.DiffSkipSimilarTo)
             .Where(s => !string.Equals(s, "public"))
             .ToHashSet();
-        this.targetSeqs = target.GetSequences(new Settings { Schema = settings.Schema }, skipSimilar: settings.DiffSkipSimilarTo)
+        this.targetSeqs = target.GetSequences(diffSettings, skipSimilar: settings.DiffSkipSimilarTo)
             .ToDictionary(t => new Seq(t.Schema, t.Name), t => t);
     }