+++ PgRoutiner/DataAccess/FilterTypes.cs
@@ -22,14 +22,16 @@ public static partial class DataAccessConnectionExtensions
                 ) sub
 
                 where
-                    (
-                        (   @schema is not null and sub.schema similar to @schema   )
-                        or
-                        (   {GetSchemaExpression("sub.schema")}  
-                    )
+                    (   @schema is null or (sub.schema similar to @schema)   )
+                    and (   @not_schema is null or sub.schema not similar to @not_schema   )
+                    and (   {GetSchemaExpression("sub.schema")}  )
+
                     and (   @skipSimilar is null or (sub.name not similar to @skipSimilar)   )
                 
-                ", ("schema", settings.Schema, DbType.AnsiString), ("skipSimilar", skipSimilar, DbType.AnsiString)).Select(t => new PgItem
+                ", 
+                ("schema", settings.SchemaSimilarTo, DbType.AnsiString),
+                ("not_schema", settings.SchemaNotSimilarTo, DbType.AnsiString),
+                ("skipSimilar", skipSimilar, DbType.AnsiString)).Select(t => new PgItem
         {
             Schema = t.Schema,
             Name = t.Name,