+++ PgRoutiner/SettingsManagement/ParseInitialSettings.cs
@@ -47,7 +47,7 @@ namespace PgRoutiner.SettingsManagement
                 }
             }
 
-            var routineCount = connection.GetRoutineCount(Value);
+            var routineCount = connection.GetRoutineCount(Value, schemaSimilarTo: Value.RoutinesSchemaSimilarTo, schemaNotSimilarTo: Value.RoutinesSchemaNotSimilarTo);
             var crudCount = connection.GetTableDefintionsCount(Value);
             if ((Value.Routines && Value.OutputDir != null && routineCount > 0) ||
                 (Value.Crud && Value.CrudOutputDir != null && crudCount > 0) ||